Move partybid tests into folder
