+++ test/PartyBid/Bid.test.js
@@ -4,11 +4,11 @@ const { waffle } = require('hardhat');
 const { provider } = waffle;
 const { expect } = require('chai');
 // ============ Internal Imports ============
-const { eth, contribute, bidThroughParty } = require('./helpers/utils');
-const { placeBid } = require('./helpers/externalTransactions');
-const { deployTestContractSetup } = require('./helpers/deploy');
-const { MARKETS, MARKET_NAMES } = require('./helpers/constants');
-const { testCases } = require('./testCases.json');
+const { eth, contribute, bidThroughParty } = require('../helpers/utils');
+const { placeBid } = require('../helpers/externalTransactions');
+const { deployTestContractSetup } = require('../helpers/deploy');
+const { MARKETS, MARKET_NAMES } = require('../helpers/constants');
+const { testCases } = require('../partyBidTestCases.json');
 
 describe('Bid', async () => {
   MARKETS.map((marketName) => {
